Added Double support for data + custom pie colors

+++ Sources/SwiftUICharts/LineChart/LineChartView.swift
@@ -17,7 +17,7 @@ public struct LineChartView: View {
     public var formSize:CGSize
     @State private var touchLocation:CGPoint = .zero
     @State private var showIndicatorDot: Bool = false
-    @State private var currentValue: Int = 2 {
+    @State private var currentValue: Double = 2 {
         didSet{
             if (oldValue != self.currentValue && showIndicatorDot) {
 //                selectionFeedbackGenerator.selectionChanged()
@@ -28,7 +28,7 @@ public struct LineChartView: View {
     }
     let frame = CGSize(width: 180, height: 120)
     
-    public init(data: [Int], title: String, legend: String? = nil, style: ChartStyle = Styles.lineChartStyleOne, form: CGSize? = Form.medium){
+    public init(data: [Double], title: String, legend: String? = nil, style: ChartStyle = Styles.lineChartStyleOne, form: CGSize? = Form.medium){
         self.data = ChartData(points: data)
         self.title = title
         self.legend = legend
